Safely Refactored what was originally the product component to the home page
Mistake on our part to not originally name it the home page as that was essentially what it was. But all fixed.
