--- src/Components/Products Page/index.jsx
@@ -1,9 +1,0 @@
-
-const ProductsPage = () => {
-    return (
-        <>
-        <h2>ProductsPage</h2>
-        </>
-    )
-}
-export default ProductsPage
